🚚 Move commands arbitrary into /arbitrary
