+++ src/arbitrary/_internals/CommandsArbitrary.ts
@@ -1,19 +1,16 @@
-import { Random } from '../../../random/generator/Random';
-import { Stream } from '../../../stream/Stream';
-import { Arbitrary } from '../../arbitrary/definition/Arbitrary';
-import { oneof } from '../../../arbitrary/oneof';
-import { AsyncCommand } from '../command/AsyncCommand';
-import { Command } from '../command/Command';
-import { ICommand } from '../command/ICommand';
-import { ReplayPath } from '../ReplayPath';
-import { CommandsIterable } from './CommandsIterable';
-import { CommandsContraints } from './CommandsContraints';
-import { CommandWrapper } from './CommandWrapper';
-import { makeLazy } from '../../../stream/LazyIterableIterator';
-import { NextArbitrary } from '../../arbitrary/definition/NextArbitrary';
-import { IntegerArbitrary } from '../../../arbitrary/_internals/IntegerArbitrary';
-import { convertFromNext, convertToNext } from '../../arbitrary/definition/Converters';
-import { NextValue } from '../../arbitrary/definition/NextValue';
+import { Arbitrary } from '../../check/arbitrary/definition/Arbitrary';
+import { convertToNext } from '../../check/arbitrary/definition/Converters';
+import { NextArbitrary } from '../../check/arbitrary/definition/NextArbitrary';
+import { NextValue } from '../../check/arbitrary/definition/NextValue';
+import { ICommand } from '../../check/model/command/ICommand';
+import { CommandsIterable } from '../../check/model/commands/CommandsIterable';
+import { CommandWrapper } from '../../check/model/commands/CommandWrapper';
+import { ReplayPath } from '../../check/model/ReplayPath';
+import { Random } from '../../random/generator/Random';
+import { makeLazy } from '../../stream/LazyIterableIterator';
+import { Stream } from '../../stream/Stream';
+import { oneof } from '../oneof';
+import { IntegerArbitrary } from './IntegerArbitrary';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 type CommandsArbitraryContext<Model extends object, Real, RunResult, CheckAsync extends boolean> = {
@@ -23,7 +20,7 @@ type CommandsArbitraryContext<Model extends object, Real, RunResult, CheckAsync 
 
 /** @internal */
 // eslint-disable-next-line @typescript-eslint/ban-types
-class CommandsArbitrary<Model extends object, Real, RunResult, CheckAsync extends boolean> extends NextArbitrary<
+export class CommandsArbitrary<Model extends object, Real, RunResult, CheckAsync extends boolean> extends NextArbitrary<
   CommandsIterable<Model, Real, RunResult, CheckAsync>
 > {
   readonly oneCommandArb: NextArbitrary<CommandWrapper<Model, Real, RunResult, CheckAsync>>;
@@ -173,98 +170,3 @@ class CommandsArbitrary<Model extends object, Real, RunResult, CheckAsync extend
     });
   }
 }
-
-/**
- * For arrays of {@link AsyncCommand} to be executed by {@link asyncModelRun}
- *
- * This implementation comes with a shrinker adapted for commands.
- * It should shrink more efficiently than {@link array} for {@link AsyncCommand} arrays.
- *
- * @param commandArbs - Arbitraries responsible to build commands
- * @param maxCommands - Maximal number of commands to build
- *
- * @deprecated
- * Superceded by `fc.commands(commandArbs, {maxCommands})` - see {@link https://github.com/dubzzz/fast-check/issues/992 | #992}.
- * Ease the migration with {@link https://github.com/dubzzz/fast-check/tree/main/codemods/unify-signatures | our codemod script}.
- *
- * @remarks Since 1.5.0
- * @public
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-function commands<Model extends object, Real, CheckAsync extends boolean>(
-  commandArbs: Arbitrary<AsyncCommand<Model, Real, CheckAsync>>[],
-  maxCommands?: number
-): Arbitrary<Iterable<AsyncCommand<Model, Real, CheckAsync>>>;
-/**
- * For arrays of {@link Command} to be executed by {@link modelRun}
- *
- * This implementation comes with a shrinker adapted for commands.
- * It should shrink more efficiently than {@link array} for {@link Command} arrays.
- *
- * @param commandArbs - Arbitraries responsible to build commands
- * @param maxCommands - Maximal number of commands to build
- *
- * @deprecated
- * Superceded by `fc.commands(commandArbs, {maxCommands})` - see {@link https://github.com/dubzzz/fast-check/issues/992 | #992}.
- * Ease the migration with {@link https://github.com/dubzzz/fast-check/tree/main/codemods/unify-signatures | our codemod script}.
- *
- * @remarks Since 1.5.0
- * @public
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-function commands<Model extends object, Real>(
-  commandArbs: Arbitrary<Command<Model, Real>>[],
-  maxCommands?: number
-): Arbitrary<Iterable<Command<Model, Real>>>;
-/**
- * For arrays of {@link AsyncCommand} to be executed by {@link asyncModelRun}
- *
- * This implementation comes with a shrinker adapted for commands.
- * It should shrink more efficiently than {@link array} for {@link AsyncCommand} arrays.
- *
- * @param commandArbs - Arbitraries responsible to build commands
- * @param constraints - Contraints to be applied when generating the commands
- *
- * @remarks Since 1.11.0
- * @public
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-function commands<Model extends object, Real, CheckAsync extends boolean>(
-  commandArbs: Arbitrary<AsyncCommand<Model, Real, CheckAsync>>[],
-  constraints?: CommandsContraints
-): Arbitrary<Iterable<AsyncCommand<Model, Real, CheckAsync>>>;
-/**
- * For arrays of {@link Command} to be executed by {@link modelRun}
- *
- * This implementation comes with a shrinker adapted for commands.
- * It should shrink more efficiently than {@link array} for {@link Command} arrays.
- *
- * @param commandArbs - Arbitraries responsible to build commands
- * @param constraints - Constraints to be applied when generating the commands
- *
- * @remarks Since 1.11.0
- * @public
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-function commands<Model extends object, Real>(
-  commandArbs: Arbitrary<Command<Model, Real>>[],
-  constraints?: CommandsContraints
-): Arbitrary<Iterable<Command<Model, Real>>>;
-// eslint-disable-next-line @typescript-eslint/ban-types
-function commands<Model extends object, Real, RunResult, CheckAsync extends boolean>(
-  commandArbs: Arbitrary<ICommand<Model, Real, RunResult, CheckAsync>>[],
-  constraints?: number | CommandsContraints
-): Arbitrary<Iterable<ICommand<Model, Real, RunResult, CheckAsync>>> {
-  const config =
-    constraints == null ? {} : typeof constraints === 'number' ? { maxCommands: constraints } : constraints;
-  return convertFromNext(
-    new CommandsArbitrary(
-      commandArbs,
-      config.maxCommands != null ? config.maxCommands : 10,
-      config.replayPath != null ? config.replayPath : null,
-      !!config.disableReplayLog
-    )
-  );
-}
-
-export { commands };
