+++ test/unit/check/model/commands/CommandsArbitrary.spec.ts
@@ -4,7 +4,7 @@ import * as fc from '../../../../../lib/fast-check';
 import { Command } from '../../../../../src/check/model/command/Command';
 import { Random } from '../../../../../src/random/generator/Random';
 import { constant } from '../../../../../src/arbitrary/constant';
-import { commands } from '../../../../../src/check/model/commands/CommandsArbitrary';
+import { commands } from '../../../../../src/arbitrary/commands';
 import { genericTuple } from '../../../../../src/arbitrary/genericTuple';
 import { nat } from '../../../../../src/arbitrary/nat';
 import { Arbitrary } from '../../../../../src/check/arbitrary/definition/Arbitrary';
